revert

+++ internal/engine/interpreter/interpreter.go
@@ -1358,7 +1358,7 @@ func (ce *callEngine) callNativeFunc(ctx context.Context, callCtx *wasm.CallCont
 				ce.pushValue(uint64(uint32(v2) ^ uint32(v1)))
 			} else {
 				// UnsignedInt64
-				ce.pushValue(v2 ^ v1)
+				ce.pushValue(uint64(v2 ^ v1))
 			}
 			frame.pc++
 		case wazeroir.OperationKindShl:
@@ -1405,7 +1405,7 @@ func (ce *callEngine) callNativeFunc(ctx context.Context, callCtx *wasm.CallCont
 				ce.pushValue(uint64(bits.RotateLeft32(uint32(v1), -int(v2))))
 			} else {
 				// UnsignedInt64
-				ce.pushValue(bits.RotateLeft64(v1, -int(v2)))
+				ce.pushValue(uint64(bits.RotateLeft64(v1, -int(v2))))
 			}
 			frame.pc++
 		case wazeroir.OperationKindAbs:
